requirements.txt
# Job Discovery System Requirements

# Core
python>=3.10

# Configuration
pyyaml>=6.0
python-dotenv>=0.19.0

# Web scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0  # For career page scraping if needed

# Web dashboard
flask>=3.0.0

# Email
jinja2>=3.1.0

# Data processing
pandas>=1.3.0
python-dateutil>=2.8.0

# Logging
python-json-logger>=2.0.0

# API clients
linkedin-api>=2.0.0  # LinkedIn API client (if available)

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0

# Development
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
